pyproject.toml
[tool.poetry]
name = "fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sentence-transformers = "^2.3.1"
elasticsearch = "^8.12.0"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.1"
pydantic = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
